feat: penumbra-operator
First draft of a k8s operator, written in Rust, for deploying Penumbra
nodes. Prepared this as an exercise to learn the k8s internals more.
Doesn't handle provisioning external LBs for public IPs yet, but still
does a lot:

  * defines CRD "PenumbraNode"
  * creates statefulset and service for node
  * maps changes to CRD to resources via reconcile loop

The Rust code is rather verbose in declaring the k8s resources, mostly
via liberal sprinkling of `Default::default()`.

The error handling is cribbed from the `runo` operator. AFAICT,
use of anyhow is not possible, because `anyhow::Error` doesn't impl
`std::error::Error`. Switching to `thiserror` seems to resolve.

feat: tf config for integration testing

Creating a cluster is easy enough, as is running the operator within it.
Creating a CRD is honored by the API, but no actual resources are
created based on that. Why not? The logged error is pretty
uninformative.

chore: update penumbra version vars

fix: support archive_url

fix: doks creds refresh

Needed to update the envrc files to auth against digitalocean.
This is a clumsy approach: we're clobbering ~/.kube/config
each time we change clusters. That's OK bc it ensures we've got
scripting in place to auth ad-hoc.

For the do-k8s cluster, the auth tokens expire quickly, which is
annoying. Tried without `--delay-seconds`, as the docs indicate
that'd result in permanent tokens, but that didn't seem to be the case.

build(penumbra-operator): add containerfile