feat: actually support event indexing
Hooked up a few needed hoses. Now, the pg container and db is created,
with the schema initialized, but the cometbft config isn't updated on
the fly.

TODO: add an env var for the init container that includes the hash of
the script.

Tried to set enable_indexing=true by default, but doesn't look like it's
taking effect.